app/requirements.txt
aiofiles==0.5.0
amqp==5.0.6
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
billiard==3.6.4.0
black==21.9b0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.1
flake8==3.9.2
flower==1.0.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.2
humanize==3.11.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
motor==2.5.1
mypy==0.910
mypy-extensions==0.4.3
orjson==3.6.3
pathspec==0.9.0
platformdirs==2.3.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
pycodestyle==2.7.0
pydantic==1.8.2
pyflakes==2.3.1
pymongo==3.12.0
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.28
requests==2.26.0
Rx==1.6.1
six==1.16.0
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
tornado==6.1
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.13.4
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
